refactor: 라우팅 경로 변경 및 스타일 변경
