+++ app/posts/[postId]/components/clientside/index.ts
@@ -1,0 +1,15 @@
+import CommentBox from "./commentBox";
+import CommentMenu from "./commentMenu";
+import UserCommentForm from "./userCommentForm";
+import TOC from "./TOC";
+
+const Components = {
+  CommentBox,
+  CommentMenu,
+  UserCommentForm,
+  TOC,
+};
+
+export default Components;
+
+export { CommentBox, CommentMenu, UserCommentForm, TOC };
