+++ app/posts/[postId]/page.tsx
@@ -1,8 +1,8 @@
-import CommentBox from "./components/clientside/commentBox";
 import PostNavigate from "./components/postNavigate";
 import HashTag from "./components/hashTag";
 import SeriesBoard from "app/components/clientside/seriesBoard";
 import styles from "./styles/page.module.scss";
+import { CommentBox } from "./components/clientside";
 import { getPost, mdParser } from "./utils";
 import { CardLayout } from "app/components/card";
